Only create the pidfile if we're daemonizing

+++ hubblestack/daemon.py
@@ -47,8 +47,8 @@ def run():
 
     if __opts__['daemonize']:
         salt.utils.daemonize()
+        create_pidfile()
 
-    create_pidfile()
     signal.signal(signal.SIGTERM, clean_up_process)
     signal.signal(signal.SIGINT, clean_up_process)
 
@@ -368,6 +368,7 @@ def clean_up_process(signal, frame):
     '''
     Clean up pidfile and anything else that needs to be cleaned up
     '''
-    if os.path.isfile(__opts__['pidfile']):
-        os.remove(__opts__['pidfile'])
+    if __opts__['daemonize']:
+        if os.path.isfile(__opts__['pidfile']):
+            os.remove(__opts__['pidfile'])
     sys.exit(0)
